refactor: change path of file navigation.ts
